9-gram solver code added
added code to load and use custom "folded" 9-gram file format which aliases the location of 4 9-grams into each cell and then keys all the data to indicate which one is present.

also adds in some enhancements to the letter picking code for 8-gram and 10-gram files too. essentially tries to pick a location where the letter picker will find an optimization surface to travel on rather than attempting blind changes in the nosiest parts of a cipher.