new 8-grams and 9-grams
these are just the config files and table files. full 8-gram and 9-gram files need to be downloaded from the links added to the main README.md file

Have also update the Readme.txt in N-grams folder to document the 8-gram format. The 9-gram format is more complicated and involves aliasing parts of the data table on top of itself 4x and then keying the log values to distinguish which are which. This results in lower dynamic range [0,66] than other formats that span [0,255] but it's still sufficient to test for research purposes.